Server and client security and error handling improvements

Closes #17, #21, and #5

See merge request it2810-h18/prosjekt3/gruppe45!8

+++ server/src/index.js
@@ -1,17 +1,27 @@
 import express from 'express';
+import { check, validationResult } from 'express-validator/check';
 import bodyParser from 'body-parser';
 import cors from 'cors';
 import { Song } from './models';
 import { recomputeIndex, search } from './search';
 
+// The number of songs to show on a single page.
 const PAGE_SIZE = 2;
 
 const app = express();
 
+// Allow cross-origin resource sharing.
 app.use(cors());
 
 // Used for parsing JSON data using Express.
 app.use(bodyParser.json());
+
+// Handle errors. This is a middleware which needs to
+// be added last, after all other middlewares.
+app.use(function (err, req, res, next) {
+  console.error(err.stack);
+  res.status(500).json();
+})
 
 // Either run on the specified port, or default to 8000.
 const port = process.env.PORT || 8000;
@@ -42,14 +52,31 @@ initialize().then(() => {
 });
 
 // Create a song
-app.post('/songs', async (req, res) => {
-  const model = await Song.create({
-    title: req.body.title,
-    artist: req.body.artist,
-    album: req.body.album,
-    genre: req.body.genre,
-    description: req.body.description,
-  });
+app.post('/songs', [
+  check('title').isString(),
+  check('artist').isString(),
+  check('album').isString(),
+  check('genre').isString(),
+  check('description').isString(),
+], async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
+  let model;
+
+  try {
+    model = await Song.create({
+      title: req.body.title,
+      artist: req.body.artist,
+      album: req.body.album,
+      genre: req.body.genre,
+      description: req.body.description,
+    });
+  } catch (error) {
+    next(error);
+  }
 
   // Recompute the search index in the background.
   recomputeIndex();
@@ -58,28 +85,43 @@ app.post('/songs', async (req, res) => {
 });
 
 // Get all songs
-app.get('/songs', async (req, res) => {
+app.get('/songs', [
+  check('page').isInt({
+    min: 0,
+    allow_leading_zeroes: false,
+  }),
+  check('search').isString().optional(),
+], async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
   const page = parseInt(req.query.page, 10);
 
   const searchQuery = req.query.search;
 
   let songs;
 
-  if (searchQuery) {
-    const result = search(
-      // Strip all non-alphanumeric and space characters from the query, as
-      // these are not handled by the search engine correctly.
-      // See https://stackoverflow.com/questions/6053541/regex-every-non-alphanumeric-character-except-white-space-or-colon
-      searchQuery.replace(/[^a-zA-Z\d\s]/g, ''),
-    ).map(result => result.ref);
+  try {
+    if (searchQuery) {
+      const result = search(
+        // Strip all non-alphanumeric and space characters from the query, as
+        // these are not handled by the search engine correctly.
+        // See https://stackoverflow.com/questions/6053541/regex-every-non-alphanumeric-character-except-white-space-or-colon
+        searchQuery.replace(/[^a-zA-Z\d\s]/g, ''),
+      ).map(result => result.ref);
 
-    songs = await Song.findAll({
-      where: {
-        id: result,
-      },
-    });
-  } else {
-    songs = await Song.findAll();
+      songs = await Song.findAll({
+        where: {
+          id: result,
+        },
+      });
+    } else {
+      songs = await Song.findAll();
+    }
+  } catch (error) {
+    next(error);
   }
 
   const result = {
@@ -91,14 +133,60 @@ app.get('/songs', async (req, res) => {
 });
 
 // Get a specific song
-app.get('/songs/:id', async (req, res) => {
-  const song = await Song.findByPk(req.params.id);
-  res.status(200).json(song);
+app.get('/songs/:id', [
+  check('id').isInt({
+    min: 0,
+    allow_leading_zeroes: false,
+  }),
+], async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
+  let song;
+
+  try {
+    song = await Song.findByPk(req.params.id);
+  } catch (error) {
+    next(error);
+  }
+
+  if (song) {
+    res.status(200).json(song);
+  } else {
+    res.status(404).json();
+  }
 });
 
 // Update a specific song
-app.put('/songs/:id', async (req, res) => {
-  const song = await Song.findByPk(req.params.id);
+app.put('/songs/:id', [
+  check('id').isInt({
+    min: 0,
+    allow_leading_zeroes: false,
+  }),
+  check('title').isString().optional(),
+  check('artist').isString().optional(),
+  check('album').isString().optional(),
+  check('genre').isString().optional(),
+  check('description').isString().optional(),
+], async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
+  let song;
+  
+  try {
+    song = await Song.findByPk(req.params.id);
+  } catch (error) {
+    next(error);
+  }
+  
+  if (!song) {
+    return res.status(422).json();
+  }
 
   // Merge the provided data with the old song.
   const updated = {
@@ -109,9 +197,13 @@ app.put('/songs/:id', async (req, res) => {
     description: req.body.description || song.dataValues.description,
   };
 
-  await Song.update(updated, {
-    where: { id: req.params.id },
-  });
+  try {
+    await Song.update(updated, {
+      where: { id: req.params.id },
+    });
+  } catch (error) {
+    next(error);
+  }
 
   // Recompute the index in the background.
   recomputeIndex();
@@ -120,10 +212,24 @@ app.put('/songs/:id', async (req, res) => {
 });
 
 // Delete a specific song
-app.delete('/songs/:id', async (req, res) => {
-  await Song.destroy({
-    where: { id: req.params.id },
-  });
+app.delete('/songs/:id', [
+  check('id').isInt({
+    min: 0,
+    allow_leading_zeroes: false,
+  }),
+], async (req, res, next) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return res.status(422).json({ errors: errors.array() });
+  }
+
+  try {
+    await Song.destroy({
+      where: { id: req.params.id },
+    });
+  } catch (error) {
+    next(error);
+  }
 
   // Recompute the index in the background.
   recomputeIndex();
